% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyLoadR.R
\name{plyLoadR}
\alias{plyLoadR}
\title{Load ply files into a three.js scene}
\usage{
plyLoadR(
  paths,
  localFiles = TRUE,
  plyCopiesFolder = "ply_local_copies",
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{plyFiles}{A list of ply files to load.}
}
\description{
Load ply files into a three.js scene
}
