% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyLoadR.R
\name{plyLoadR}
\alias{plyLoadR}
\title{Load ply files into a three.js scene}
\usage{
plyLoadR(
  paths,
  localFiles = TRUE,
  plyCopiesFolder = "ply_local_copies",
  isWireframe = FALSE,
  isTransparent = FALSE,
  opacity = 1,
  camera = NULL,
  controls = NULL,
  elementId = NULL,
  toggleMeshes = NULL,
  showLoadingProgress = FALSE,
  width = "100\%",
  height = "65vh"
)
}
\arguments{
\item{paths}{Paths of the ply files to load.}

\item{localFiles}{Boolean.}
}
\description{
Load ply files into a three.js scene
}
